add router for recipe list view
